requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==4.2
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.13.1
django-appconf==1.0.6
django-better-admin-arrayfield==1.4.2
django-ckeditor==6.7.2
django-cors-headers==3.13.0
django-debug-toolbar==3.2.4
django-import-export==4.1.1
django-js-asset==2.1.0
django-multi-captcha-admin==2.0.0
django-nested-admin==4.1.1
django-recaptcha2==1.4.1
django-select2==8.2.1
django-simple-history==3.8.0
django-subadmin==3.2.4
django-unfold==0.22.0
djangorestframework==3.16.0
et-xmlfile==1.1.0
gunicorn==23.0.0
idna==3.8
MarkupPy==1.14
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
